backend/reservation-service/requirements.txt
Flask==2.3.2
Flask-Session==0.8.0
Werkzeug==2.3.7
PyMySQL==1.1.0
python-dotenv==1.0.0
gunicorn 
Flask-Cors
mysql-connector-python